+++ tailwind.config.js
@@ -1,8 +1,3 @@
-module.exports = {
-  content: [
-    './app/views/**/*.html.erb',
-    './app/helpers/**/*.rb',
-    './app/assets/stylesheets/**/*.css',
-    './app/javascript/**/*.js'
-  ]
-}
+export default {
+  plugins: ["@tailwindcss/postcss"],
+};